pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.2"
description = "Compares two configuration files and shows a difference."
authors = ["shakedizzy <shakedizzy81@gmail.com>"]
packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
